flytt /sykmelding til /sykmeldinger
